requirements.txt
tensorflow-gpu==2.4.0
tensorflow-datasets
matplotlib
hickle
tqdm